fix: move handleJson to plugin
